Extracted eventuate-common-reactive-jdbc from eventuate-common-spring-reactive-jdbc (necessary for https://github.com/eventuate-tram/eventuate-tram-core/issues/143)
