+++ eventuate-common-spring-reactive-jdbc/build.gradle
@@ -1,9 +1,9 @@
 
 dependencies {
+    compile project (":eventuate-common-reactive-jdbc")
     compile project (":eventuate-common-common-spring-jdbc")
     compile project (":eventuate-common-spring-id")
 
-    compile "dev.miku:r2dbc-mysql:$r2dbcMysqlVersion"
     compile "org.springframework.data:spring-data-r2dbc:$springDataR2dbcVersion"
 
     testCompile project (":eventuate-common-spring-jdbc")
